support lollypop

+++ xfce-spotify-main.sh
@@ -3,7 +3,7 @@
 # Script orignally made by macr1408 (https://github.com/macr1408), modified by userbyte (https://github.com/userbyte) to use playerctl instead of the Spotify API because I could't get it to work
 # Made for non-commercial use
 
-if ! pgrep -x spotify >/dev/null && ! pgrep -x chrome >/dev/null && ! pgrep -x firefox >/dev/null
+if ! pgrep -x spotify >/dev/null && ! pgrep -x chrome >/dev/null && ! pgrep -x firefox >/dev/null && ! pgrep -f lollypop >/dev/null
 then
     echo "<txt></txt>"
     exit 1;
